+++ examples/snippets/src/print_vm_addresses.rs
@@ -77,5 +77,6 @@ async fn main() -> anyhow::Result<()> {
             vm.name
         );
     }
+    client.close().await?;
     Ok(())
 }