+++ examples/snippets/src/retrieve_ds_hosts.rs
@@ -57,5 +57,6 @@ async fn main() -> anyhow::Result<()> {
         );
     }
 
+    client.close().await?;
     Ok(())
 }